requirements.txt
beautifulsoup4==4.12.2
discord.py==2.2.3
feedparser==6.0.10
ndjson==0.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
scrapetube==2.3.1
soupsieve==2.4.1
